% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianw_f.R
\name{meanw_f}
\alias{meanw_f}
\title{Mean Weibull survival function}
\usage{
meanw_f(ascale, bshape)
}
\arguments{
\item{ascale}{scale parameter for the Weibull distribution}

\item{bshape}{shape parameter for the Weibull distribution}
}
\value{
mean
}
\description{
The functions `meanw_f` and `medianw_f` calculate the mean and median for Weibull distributions, respectively.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
