% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianw_f.R
\name{medianw_f}
\alias{medianw_f}
\title{Median Weibull survival function}
\usage{
medianw_f(ascale, bshape)
}
\arguments{
\item{ascale}{scale parameter for the Weibull distribution}

\item{bshape}{shape parameter for the Weibull distribution}
}
\value{
median
}
\description{
The functions `meanw_f` and `medianw_f` calculate the mean and median for Weibull distributions, respectively.
}
\author{
Marta Bofill Roig

The functions `meanw_f` and `medianw_f` calculate the mean and median for Weibull distributions, respectively.
}
\keyword{internal}
