% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scale.R
\name{param_scale}
\alias{param_scale}
\title{Scale parameter computation}
\usage{
param_scale(s, t, shape = 1)
}
\arguments{
\item{s}{survival rate at time t}

\item{t}{time at which the survival distribution is evaluated}

\item{shape}{shape parameter for the Weibull distribution}
}
\value{
Variance computation
}
\description{
returns the value of the scale parameter a given the survival (s) at time t
}
\note{
Weibull parametrization: S(x) =  exp(- (x/a)^b)
}
\author{
Marta Bofill Roig
}
\keyword{internal}
