% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale1_taylorf.R
\name{scale1_taylorf}
\alias{scale1_taylorf}
\title{Scale parameter computation}
\usage{
scale1_taylorf(ascale0, Delta, tau)
}
\arguments{
\item{ascale0}{scale parameter for the weibull distribution in the control group}

\item{Delta}{RMST difference between groups}

\item{tau}{end of follow-up}
}
\value{
Variance computation
}
\description{
returns the value of the scale parameter in the intervention group using Taylor series
}
\note{
Weibull parametrization: S(x) =  exp(- (x/a)^b)
}
\author{
Marta Bofill Roig
}
\keyword{internal}
