% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survw_derivf.R
\name{survw_derivf}
\alias{survw_derivf}
\title{Derivative Weibull survival function}
\usage{
survw_derivf(t, ascale, bshape = 1)
}
\arguments{
\item{t}{time}

\item{ascale}{scale parameter for the Weibull distribution}

\item{bshape}{shape parameter for the Weibull distribution}
}
\value{
derivative
}
\description{
The function `survw_derivf` computes the derivative of the survival distribution  `survw_f`.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
