% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survw_f.R
\name{survw_f}
\alias{survw_f}
\title{Weibull survival function}
\usage{
survw_f(t, ascale, bshape)
}
\arguments{
\item{t}{time}

\item{ascale}{scale parameter for the Weibull distribution}

\item{bshape}{shape parameter for the Weibull distribution}
}
\value{
survival function
}
\description{
The function `survw_f` computes the Weibull survival function.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
