% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_f.R
\name{survw_integratef}
\alias{survw_integratef}
\title{Integrate function}
\usage{
survw_integratef(t, tau, ascale, bshape)
}
\arguments{
\item{t}{time}

\item{tau}{follow-up}

\item{ascale}{scale parameter for the Weibull distribution}

\item{bshape}{shape parameter for the Weibull distribution}
}
\value{
Variance computation
}
\description{
the function `survw_integratef` is used for the integrations
}
\author{
Marta Bofill Roig
}
\keyword{internal}
