% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_f.R
\name{var_f}
\alias{var_f}
\title{Variance computation}
\usage{
var_f(ascale_r, ascale_nr, tau, bshape, ascale_cens, p)
}
\arguments{
\item{ascale_r}{scale parameter for the Weibull distribution   for responders}

\item{ascale_nr}{scale parameter for the Weibull distribution  for non-responders}

\item{tau}{follow-up}

\item{bshape}{shape parameter for the Weibull distribution}

\item{ascale_cens}{distributional parameter for the exponential distribution for the censoring}

\item{p}{event rate for the response}
}
\value{
Variance computation
}
\description{
The   function  `var_f` computes the variance.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
