% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_f.R
\name{inside_var}
\alias{inside_var}
\title{Inside variance computation}
\usage{
inside_var(t, lambda_r, lambda_nr, tau, bet, lambda_cens, p)
}
\arguments{
\item{t}{time at which the survival distribution is evaluated}

\item{lambda_r}{scale parameter for the Weibull distribution   for responders}

\item{lambda_nr}{scale parameter for the Weibull distribution  for non-responders}

\item{tau}{follow-up}

\item{bet}{shape parameter for the Weibull distribution}

\item{lambda_cens}{distributional parameter for the exponential distribution for the censoring}

\item{p}{event rate for the response}
}
\value{
Variance computation
}
\description{
The following three functions are used to calculate the variance of th difference of two RMSTs. `survw_integratef` is used for the integrations; `inside_var` calculates the expression inside the integral; finally, `var_f` computes the variance.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
