% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstw_f.R
\name{rmstw_f}
\alias{rmstw_f}
\title{Restricted mean survival times Weibull distribution}
\usage{
rmstw_f(lambda, bet, tau, low = 0)
}
\arguments{
\item{lambda}{scale parameter for the Weibull distribution}

\item{bet}{shape parameter for the Weibull distribution}

\item{tau}{rmst evaluated from low to tau}

\item{low}{rmst evaluated from low to tau}
}
\value{
rmst
}
\description{
The function `rmstw_f` computes the restricted mean survival times (RMST) according to the Weibull survival function.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
