% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survmixture_f.R
\name{survmixture_f}
\alias{survmixture_f}
\title{Mixture survival function}
\usage{
survmixture_f(t, lambda_r, lambda_nr, bet = 1, p)
}
\arguments{
\item{t}{time at which the survival distribution is evaluated}

\item{lambda_r}{scale parameter for the Weibull distribution   for responders}

\item{lambda_nr}{scale parameter for the Weibull distribution  for non-responders}

\item{bet}{shape parameter for the Weibull distribution}

\item{p}{event rate for the response}
}
\value{
Mixture survival function evaluated at t
}
\description{
The function `survmixture_f` computes the survival distribution as a mixture of  of responders and non-responders. The responders and non-responders distributions are assumed to be Weibull distributions.
}
\author{
Marta Bofill Roig
}
