% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survw_effectsize.R
\name{survw_effectsize}
\alias{survw_effectsize}
\title{Effect size calculation for mixture survival distributions}
\usage{
survw_effectsize(
  ascale0_r,
  ascale0_nr,
  delta_p,
  p0,
  bshape0,
  bshape1,
  ascale1_r,
  ascale1_nr,
  tau,
  Delta_r = NULL,
  Delta_0 = NULL,
  Delta_nr = NULL,
  anticipated_effects = FALSE
)
}
\arguments{
\item{ascale0_r}{scale parameter for the Weibull distribution in the control group for responders}

\item{ascale0_nr}{scale parameter for the Weibull distribution in the control group for non-responders}

\item{delta_p}{effect size for the response rate}

\item{p0}{event rate for the response}

\item{bshape0}{shape parameter for the Weibull distribution in the control group}

\item{bshape1}{shape parameter for the Weibull distribution in the intervention group}

\item{ascale1_r}{scale parameter for the Weibull distribution in the intervention group for responders}

\item{ascale1_nr}{scale parameter for the Weibull distribution in the intervention group for non-responders}

\item{tau}{follow-up}

\item{Delta_r}{RMST difference between intervention and control groups for responders}

\item{Delta_0}{RMST difference between responders and non-responders in the control group}

\item{Delta_nr}{RMST difference between intervention and control groups for non-responders}

\item{anticipated_effects}{Logical parameter. If it is TRUE then the effect size is computed based on previous information on the effect sizes on response rate and survival-by-responses (that is, based on Delta_r, Delta_0, Delta_nr); otherwise is based on the distributional parameters (ascale0_r, ascale0_nr, ascale1_r, ascale1_nr, bshape0, bshape1).}
}
\value{
Effect size for overall survival
}
\description{
The function `survw_effectsize` calculates the effect size in terms of the difference of restricted mean survival times (RMST) according to the information on responders and non-responders.
}
\author{
Marta Bofill Roig
}
