% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survw_samplesize.R
\name{survw_samplesize}
\alias{survw_samplesize}
\title{Sample size calculation for mixture survival distributions}
\usage{
survw_samplesize(
  lambda0_r,
  lambda0_nr,
  delta_p,
  p0,
  beta0,
  beta1,
  lambda1_r,
  lambda1_nr,
  lambda_cens,
  tau,
  alpha = 0.025,
  beta = 0.2
)
}
\arguments{
\item{lambda0_r}{scale parameter for the Weibull distribution in the control group for responders}

\item{lambda0_nr}{scale parameter for the Weibull distribution in the control group for non-responders}

\item{delta_p}{effect size for the response rate}

\item{p0}{event rate for the response}

\item{beta0}{shape parameter for the Weibull distribution in the control group}

\item{beta1}{shape parameter for the Weibull distribution in the intervention group}

\item{lambda1_r}{scale parameter for the Weibull distribution in the intervention group for responders}

\item{lambda1_nr}{scale parameter for the Weibull distribution in the intervention group for non-responders}

\item{lambda_cens}{distributional parameter for the exponential distribution for the censoring}

\item{tau}{follow-up}

\item{alpha}{type I error}

\item{beta}{type II error}
}
\value{
Sample size for overall survival
}
\description{
The function `survw_samplesize` calculates the sample size according to the distributional parameters of the responders and non-responders.
}
\author{
Marta Bofill Roig
}
