% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_f.R
\name{var_f}
\alias{var_f}
\title{Variance computation}
\usage{
var_f(lambda_r, lambda_nr, tau, bet, lambda_cens, p)
}
\arguments{
\item{lambda_r}{scale parameter for the Weibull distribution   for responders}

\item{lambda_nr}{scale parameter for the Weibull distribution  for non-responders}

\item{tau}{follow-up}

\item{bet}{shape parameter for the Weibull distribution}

\item{lambda_cens}{distributional parameter for the exponential distribution for the censoring}

\item{p}{event rate for the response}
}
\value{
Variance computation
}
\description{
The   function  `var_f` computes the variance.
}
\author{
Marta Bofill Roig
}
\keyword{internal}
